Weaken test for typed must only start with the actually typed text
In some cases Google adds extra characters to what we considered the
typed text. For instance if one types "t" it suggests "t mobile" and
puts the whitespace in the typed section instead of the supplement.
Since we cannot be sure if Google only adds whitespace (it probably
does though), I we should rather not use `trim` but `startsWith` here.